+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] 
+- [Extend tests for actions](https://github.com/multiversx/mx-sdk-dapp/pull/1699)
+- [Extended tests](https://github.com/multiversx/mx-sdk-dapp/pull/1698)
+- [IframeProviderStrategy tests](https://github.com/multiversx/mx-sdk-dapp/pull/1695)
+- [Added ExtensionProviderStrategy tests](https://github.com/multiversx/mx-sdk-dapp/pull/1693)
 
 ## [[5.5.0](https://github.com/multiversx/mx-sdk-dapp/pull/1692)] - 2025-11-24